requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.2
bs4==0.0.1
CacheControl==0.12.6
certifi==2021.10.8
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
lxml==4.6.3
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.3.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
requests==2.32.3
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
tqdm==4.50.0
urllib3>=1.26
webencodings==0.5.1